+++ src/main/scala/co/topl/brambl/cli/modules/BifrostQueryModeModule.scala
@@ -19,6 +19,8 @@ trait BifrostQueryModeModule extends ChannelResourceModule {
       )
     )
     validateParams.subcmd match {
+      case BramblCliSubCmd.invalid =>
+        IO.pure(Left("A subcommand needs to be specified"))
       case BramblCliSubCmd.blockbyheight =>
         new BifrostQueryController(
           bifrostQueryAlgebra